Established tables.go and aggregates.go
