+++ sql/core.go
@@ -99,125 +99,6 @@ type NonDeterministicExpression interface {
 	IsNonDeterministic() bool
 }
 
-// Aggregation implements an aggregation expression, where an
-// aggregation buffer is created for each grouping (NewBuffer). Rows for the
-// grouping should be fed to the buffer with |Update| and the buffer should be
-// eval'd with |Eval|. Calling |Eval| directly on an Aggregation expression is
-// typically an error.
-type Aggregation interface {
-	WindowAdaptableExpression
-	// NewBuffer creates a new aggregation buffer and returns it as a Row.
-	NewBuffer() (AggregationBuffer, error)
-	// WithWindow returns a version of this aggregation with the WindowDefinition given
-	WithWindow(window *WindowDefinition) (Aggregation, error)
-	// Window returns this expression's window
-	Window() *WindowDefinition
-}
-
-// WindowBuffer is a type alias for a window materialization
-type WindowBuffer []Row
-
-// WindowInterval is a WindowBuffer index range, where [Start] is inclusive, and [End] is exclusive
-type WindowInterval struct {
-	Start, End int
-}
-
-// WindowFunction performs aggregations on buffer intervals, optionally maintaining internal state
-// for performance optimizations
-type WindowFunction interface {
-	Disposable
-
-	// WithWindow passes fields from the parent WindowDefinition, deferring partial construction of a WindowFunction
-	WithWindow(w *WindowDefinition) (WindowFunction, error)
-	// StartPartition discards any previous state and initializes the aggregation for a new partition
-	StartPartition(*Context, WindowInterval, WindowBuffer) error
-	// DefaultFramer returns a new instance of the default WindowFramer for a particular aggregation
-	DefaultFramer() WindowFramer
-	// NewSlidingFrameInterval is updates the function's internal aggregation state for the next
-	// Compute call using three WindowInterval: added, dropped, and current.
-	//TODO: implement sliding window interface in aggregation functions and windowBlockIter
-	//NewSlidingFrameInterval(added, dropped WindowInterval)
-	// Compute returns an aggregation result for a given interval and buffer
-	Compute(*Context, WindowInterval, WindowBuffer) interface{}
-}
-
-// WindowAdaptableExpression is an Expression that can be executed as a window aggregation
-type WindowAdaptableExpression interface {
-	Expression
-
-	// NewEvalable constructs an executable aggregation WindowFunction
-	NewWindowFunction() (WindowFunction, error)
-}
-
-// WindowFramer is responsible for tracking window frame indices for partition rows.
-// WindowFramer is aware of the framing strategy (offsets, ranges, etc),
-// and is responsible for returning a WindowInterval for each partition row.
-type WindowFramer interface {
-	// NewFramer is a prototype constructor that create a new Framer with pass-through
-	// parent arguments
-	NewFramer(WindowInterval) (WindowFramer, error)
-	// Next returns the next WindowInterval frame, or an io.EOF error after the last row
-	Next(*Context, WindowBuffer) (WindowInterval, error)
-	// FirstIdx returns the current frame start index
-	FirstIdx() int
-	// LastIdx returns the last valid index in the current frame
-	LastIdx() int
-	// Interval returns the current frame as a WindowInterval
-	Interval() (WindowInterval, error)
-	// SlidingInterval returns three WindowIntervals: the current frame, dropped range since the
-	// last frame, and added range since the last frame.
-	// TODO: implement sliding window interface in framers, windowBlockIter, and aggregation functions
-	//SlidingInterval(ctx Context) (WindowInterval, WindowInterval, WindowInterval)
-}
-
-// WindowFrame describe input bounds for an aggregation function
-// execution. A frame will only have two non-null fields for the start
-// and end bounds. A WindowFrame plan node is associated
-// with an exec WindowFramer.
-type WindowFrame interface {
-	fmt.Stringer
-
-	// NewFramer constructs an executable WindowFramer
-	NewFramer(*WindowDefinition) (WindowFramer, error)
-	// UnboundedFollowing returns whether a frame end is unbounded
-	UnboundedFollowing() bool
-	// UnboundedPreceding returns whether a frame start is unbounded
-	UnboundedPreceding() bool
-	// StartCurrentRow returns whether a frame start is CURRENT ROW
-	StartCurrentRow() bool
-	// EndCurrentRow returns whether a frame end is CURRENT ROW
-	EndCurrentRow() bool
-	// StartNFollowing returns a frame's start preceding Expression or nil
-	StartNPreceding() Expression
-	// StartNFollowing returns a frame's start following Expression or nil
-	StartNFollowing() Expression
-	// EndNPreceding returns whether a frame end preceding Expression or nil
-	EndNPreceding() Expression
-	// EndNPreceding returns whether a frame end following Expression or nil
-	EndNFollowing() Expression
-}
-
-type AggregationBuffer interface {
-	Disposable
-
-	// Eval the given buffer.
-	Eval(*Context) (interface{}, error)
-	// Update the given buffer with the given row.
-	Update(ctx *Context, row Row) error
-}
-
-// WindowAggregation implements a window aggregation expression. A WindowAggregation is similar to an Aggregation,
-// except that it returns a result row for every input row, as opposed to as single for the entire result set. A
-// WindowAggregation is expected to track its input rows in the order received, and to return the value for the row
-// index given on demand.
-type WindowAggregation interface {
-	WindowAdaptableExpression
-	// Window returns this expression's window
-	Window() *WindowDefinition
-	// WithWindow returns a version of this window aggregation with the window given
-	WithWindow(window *WindowDefinition) (WindowAggregation, error)
-}
-
 // Node is a node in the execution plan tree.
 type Node interface {
 	Resolvable
@@ -322,123 +203,10 @@ type PrimaryKeySchemaTarget interface {
 	WithPrimaryKeySchema(schema PrimaryKeySchema) (Node, error)
 }
 
-// TableFunction is a node that is generated by a function
-type TableFunction interface {
-	Node
-	Expressioner
-	Databaser
-	Nameable
-
-	// NewInstance returns a new instance of the table function
-	NewInstance(ctx *Context, db Database, expressions []Expression) (Node, error)
-}
-
-// Table represents the backend of a SQL table.
-type Table interface {
-	Nameable
-	String() string
-	Schema() Schema
-	Collation() CollationID
-	Partitions(*Context) (PartitionIter, error)
-	PartitionRows(*Context, Partition) (RowIter, error)
-}
-
-type Table2 interface {
-	Table
-
-	PartitionRows2(ctx *Context, part Partition) (RowIter2, error)
-}
-
-type TemporaryTable interface {
-	IsTemporary() bool
-}
-
-// TableWrapper is a node that wraps the real table. This is needed because
-// wrappers cannot implement some methods the table may implement.
-type TableWrapper interface {
-	// Underlying returns the underlying table.
-	Underlying() Table
-}
-
 // PartitionCounter can return the number of partitions.
 type PartitionCounter interface {
 	// PartitionCount returns the number of partitions.
 	PartitionCount(*Context) (int64, error)
-}
-
-// FilteredTable is a table that can produce a specific RowIter
-// that's more optimized given the filters.
-type FilteredTable interface {
-	Table
-	Filters() []Expression
-	HandledFilters(filters []Expression) []Expression
-	WithFilters(ctx *Context, filters []Expression) Table
-}
-
-// ProjectedTable is a table that can produce a specific RowIter
-// that's more optimized given the columns that are projected.
-type ProjectedTable interface {
-	Table
-	WithProjections(colNames []string) Table
-	Projections() []string
-}
-
-// IndexAddressable is a table that can be scanned through a primary index
-type IndexAddressable interface {
-	// IndexedAccess returns a table that can perform scans constrained to
-	// an IndexLookup on the index given
-	IndexedAccess(Index) IndexedTable
-	// GetIndexes returns an array of this table's Indexes
-	GetIndexes(ctx *Context) ([]Index, error)
-}
-
-type IndexAddressableTable interface {
-	Table
-	IndexAddressable
-}
-
-// IndexedTable is a table with an index chosen for range scans
-type IndexedTable interface {
-	Table
-	// LookupPartitions returns partitions scanned by the given IndexLookup
-	LookupPartitions(*Context, IndexLookup) (PartitionIter, error)
-}
-
-// IndexAlterableTable represents a table that supports index modification operations.
-type IndexAlterableTable interface {
-	Table
-	// CreateIndex creates an index for this table, using the provided parameters.
-	// Returns an error if the index name already exists, or an index with the same columns already exists.
-	CreateIndex(ctx *Context, indexDef IndexDef) error
-	// DropIndex removes an index from this table, if it exists.
-	// Returns an error if the removal failed or the index does not exist.
-	DropIndex(ctx *Context, indexName string) error
-	// RenameIndex renames an existing index to another name that is not already taken by another index on this table.
-	RenameIndex(ctx *Context, fromIndexName string, toIndexName string) error
-}
-
-// ForeignKeyTable is a table that can declare its foreign key constraints, as well as be referenced.
-type ForeignKeyTable interface {
-	IndexAddressableTable
-	// CreateIndexForForeignKey creates an index for this table, using the provided parameters. Indexes created through
-	// this function are specifically ones generated for use with a foreign key. Returns an error if the index name
-	// already exists, or an index on the same columns already exists.
-	CreateIndexForForeignKey(ctx *Context, indexDef IndexDef) error
-
-	// GetDeclaredForeignKeys returns the foreign key constraints that are declared by this table.
-	GetDeclaredForeignKeys(ctx *Context) ([]ForeignKeyConstraint, error)
-	// GetReferencedForeignKeys returns the foreign key constraints that are referenced by this table.
-	GetReferencedForeignKeys(ctx *Context) ([]ForeignKeyConstraint, error)
-	// AddForeignKey adds the given foreign key constraint to the table. Returns an error if the foreign key name
-	// already exists on any other table within the database.
-	AddForeignKey(ctx *Context, fk ForeignKeyConstraint) error
-	// DropForeignKey removes a foreign key from the table.
-	DropForeignKey(ctx *Context, fkName string) error
-	// UpdateForeignKey updates the given foreign key constraint. May range from updated table names to setting the
-	// IsResolved boolean.
-	UpdateForeignKey(ctx *Context, fkName string, fk ForeignKeyConstraint) error
-	// GetForeignKeyEditor returns a ForeignKeyEditor for this table.
-	GetForeignKeyEditor(ctx *Context) ForeignKeyEditor
 }
 
 // ForeignKeyEditor is a TableEditor that is addressable via IndexLookup.
@@ -447,224 +215,8 @@ type ForeignKeyEditor interface {
 	IndexAddressable
 }
 
-// CheckTable is a table that can declare its check constraints.
-type CheckTable interface {
-	Table
-	// GetChecks returns the check constraints on this table.
-	GetChecks(ctx *Context) ([]CheckDefinition, error)
-}
-
-// CheckAlterableTable represents a table that supports check constraints.
-type CheckAlterableTable interface {
-	Table
-	// CreateCheck creates an check constraint for this table, using the provided parameters.
-	// Returns an error if the constraint name already exists.
-	CreateCheck(ctx *Context, check *CheckDefinition) error
-	// DropCheck removes a check constraint from the database.
-	DropCheck(ctx *Context, chName string) error
-}
-
-// PrimaryKeyAlterableTable represents a table that supports primary key changes.
-type PrimaryKeyAlterableTable interface {
-	Table
-	// CreatePrimaryKey creates a primary key for this table, using the provided parameters.
-	// Returns an error if the new primary key set is not compatible with the current table data.
-	CreatePrimaryKey(ctx *Context, columns []IndexColumn) error
-	// DropPrimaryKey drops a primary key on a table. Returns an error if that table does not have a key.
-	DropPrimaryKey(ctx *Context) error
-}
-
-type PrimaryKeyTable interface {
-	// PrimaryKeySchema returns this table's PrimaryKeySchema
-	PrimaryKeySchema() PrimaryKeySchema
-}
-
-// TableEditor is the base interface for sub interfaces that can update rows in a
-// table during an INSERT, REPLACE, UPDATE, or DELETE statement.
-type TableEditor interface {
-	RowReplacer
-	RowUpdater
-}
-
-type EditOpenerCloser interface {
-	// StatementBegin is called before the first operation of a statement. Integrators should mark the state of the data
-	// in some way that it may be returned to in the case of an error.
-	StatementBegin(ctx *Context)
-	// DiscardChanges is called if a statement encounters an error, and all current changes since the statement beginning
-	// should be discarded.
-	DiscardChanges(ctx *Context, errorEncountered error) error
-	// StatementComplete is called after the last operation of the statement, indicating that it has successfully completed.
-	// The mark set in StatementBegin may be removed, and a new one should be created on the next StatementBegin.
-	StatementComplete(ctx *Context) error
-}
-
-type AutoIncrementEditor interface {
-	AutoIncrementSetter
-	AutoIncrementGetter
-}
-
-// InsertableTable is a table that can process insertion of new rows.
-type InsertableTable interface {
-	Table
-	// Inserter returns an Inserter for this table. The Inserter will get one call to Insert() for each row to be
-	// inserted, and will end with a call to Close() to finalize the insert operation.
-	Inserter(*Context) RowInserter
-}
-
-// RowInserter is an insert cursor that can insert one or more values to a table.
-type RowInserter interface {
-	EditOpenerCloser
-	// Insert inserts the row given, returning an error if it cannot. Insert will be called once for each row to process
-	// for the insert operation, which may involve many rows. After all rows in an operation have been processed, Close
-	// is called.
-	Insert(*Context, Row) error
-	// Close finalizes the insert operation, persisting its result.
-	Closer
-}
-
-// DeleteableTable is a table that can process the deletion of rows
-type DeletableTable interface {
-	Table
-	// Deleter returns a RowDeleter for this table. The RowDeleter will get one call to Delete for each row to be deleted,
-	// and will end with a call to Close() to finalize the delete operation.
-	Deleter(*Context) RowDeleter
-}
-
-// RowDeleter is a delete cursor that can delete one or more rows from a table.
-type RowDeleter interface {
-	EditOpenerCloser
-	// Delete deletes the given row. Returns ErrDeleteRowNotFound if the row was not found. Delete will be called once for
-	// each row to process for the delete operation, which may involve many rows. After all rows have been processed,
-	// Close is called.
-	Delete(*Context, Row) error
-	// Close finalizes the delete operation, persisting the result.
-	Closer
-}
-
-// TruncateableTable is a table that can process the deletion of all rows.
-type TruncateableTable interface {
-	Table
-	// Truncate removes all rows from the table. If the table also implements DeletableTable and it is determined that
-	// truncate would be equivalent to a DELETE which spans the entire table, then this function will be called instead.
-	// Returns the number of rows that were removed.
-	Truncate(*Context) (int, error)
-}
-
-// AutoIncrementTable is a table that supports AUTO_INCREMENT.
-// Getter and Setter methods access the table's AUTO_INCREMENT
-// sequence. These methods should only be used for tables with
-// and AUTO_INCREMENT column in their schema.
-type AutoIncrementTable interface {
-	Table
-	// GetNextAutoIncrementValue gets the next AUTO_INCREMENT value. In the case that a table with an autoincrement
-	// column is passed in a row with the autoinc column failed, the next auto increment value must
-	// update its internal state accordingly and use the insert val at runtime.
-	// Implementations are responsible for updating their state to provide the correct values.
-	GetNextAutoIncrementValue(ctx *Context, insertVal interface{}) (uint64, error)
-	// AutoIncrementSetter returns an AutoIncrementSetter.
-	AutoIncrementSetter(*Context) AutoIncrementSetter
-}
-
-var ErrNoAutoIncrementCol = fmt.Errorf("this table has no AUTO_INCREMENT columns")
-
-// AutoIncrementSetter provides support for altering a table's
-// AUTO_INCREMENT sequence, eg 'ALTER TABLE t AUTO_INCREMENT = 10;'
-type AutoIncrementSetter interface {
-	// SetAutoIncrementValue sets a new AUTO_INCREMENT value.
-	SetAutoIncrementValue(*Context, uint64) error
-	// Close finalizes the set operation, persisting the result.
-	Closer
-}
-
-type AutoIncrementGetter interface {
-	GetNextAutoIncrementValue(ctx *Context, insertVal interface{}) (uint64, error)
-	// Close finalizes the set operation, persisting the result.
-	Closer
-}
-
 type Closer interface {
 	Close(*Context) error
-}
-
-// RowReplacer is a combination of RowDeleter and RowInserter.
-type RowReplacer interface {
-	EditOpenerCloser
-	RowInserter
-	RowDeleter
-}
-
-// ReplaceableTable allows rows to be replaced through a Delete (if applicable) then Insert.
-type ReplaceableTable interface {
-	Table
-	// Replacer returns a RowReplacer for this table. The RowReplacer will have Insert and optionally Delete called once
-	// for each row, followed by a call to Close() when all rows have been processed.
-	Replacer(ctx *Context) RowReplacer
-}
-
-// UpdatableTable is a table that can process updates of existing rows via update statements.
-type UpdatableTable interface {
-	Table
-	// Updater returns a RowUpdater for this table. The RowUpdater will have Update called once for each row to be
-	// updated, followed by a call to Close() when all rows have been processed.
-	Updater(ctx *Context) RowUpdater
-}
-
-// RowUpdater is an update cursor that can update one or more rows in a table.
-type RowUpdater interface {
-	EditOpenerCloser
-	// Update the given row. Provides both the old and new rows.
-	Update(ctx *Context, old Row, new Row) error
-	// Closer finalizes the update operation, persisting the result.
-	Closer
-}
-
-// RewritableTable is an extension to Table that makes it simpler for integrators to adapt to schema changes that must
-// rewrite every row of the table. In this case, rows are streamed from the existing table in the old schema,
-// transformed / updated appropriately, and written with the new format.
-type RewritableTable interface {
-	Table
-	AlterableTable
-
-	// ShouldRewriteTable returns whether this table should be rewritten because of a schema change. The old and new
-	// versions of the schema and modified column are provided. For some operations, one or both of |oldColumn| or
-	// |newColumn| may be nil.
-	// The engine may decide to rewrite tables regardless in some cases, such as when a new non-nullable column is added.
-	ShouldRewriteTable(ctx *Context, oldSchema, newSchema PrimaryKeySchema, oldColumn, newColumn *Column) bool
-
-	// RewriteInserter returns a RowInserter for the new schema. Rows from the current table, with the old schema, will
-	// be streamed from the table and passed to this RowInserter. Implementor tables must still return rows in the
-	// current schema until the rewrite operation completes. |Close| will be called on RowInserter when all rows have
-	// been inserted.
-	RewriteInserter(ctx *Context, oldSchema, newSchema PrimaryKeySchema, oldColumn, newColumn *Column, idxCols []IndexColumn) (RowInserter, error)
-}
-
-// DatabaseProvider is a collection of Database.
-type DatabaseProvider interface {
-	// Database gets a Database from the provider.
-	Database(ctx *Context, name string) (Database, error)
-
-	// HasDatabase checks if the Database exists in the provider.
-	HasDatabase(ctx *Context, name string) bool
-
-	// AllDatabases returns a slice of all Databases in the provider.
-	AllDatabases(ctx *Context) []Database
-}
-
-type MutableDatabaseProvider interface {
-	DatabaseProvider
-
-	// CreateDatabase creates a database and adds it to the provider's collection.
-	CreateDatabase(ctx *Context, name string) error
-
-	// DropDatabase removes a database from the provider's collection.
-	DropDatabase(ctx *Context, name string) error
-}
-
-type CollatedDatabaseProvider interface {
-	MutableDatabaseProvider
-
-	// CreateCollatedDatabase creates a collated database and adds it to the provider's collection.
-	CreateCollatedDatabase(ctx *Context, name string, collation CollationID) error
 }
 
 // ExternalStoredProcedureProvider provides access to built-in stored procedures. These procedures are implemented
@@ -692,18 +244,6 @@ type FunctionProvider interface {
 type TableFunctionProvider interface {
 	// TableFunction returns the table function with the name provided, case-insensitive
 	TableFunction(ctx *Context, name string) (TableFunction, error)
-}
-
-// UnresolvedTable is a Table that is either unresolved or deferred for until an asOf resolution
-type UnresolvedTable interface {
-	Nameable
-	// Database returns the database name
-	Database() string
-	// WithAsOf returns a copy of this versioned table with its AsOf
-	// field set to the given value. Analogous to WithChildren.
-	WithAsOf(asOf Expression) (Node, error)
-	//AsOf returns this table's asof expression.
-	AsOf() Expression
 }
 
 type TransactionCharacteristic int
@@ -755,20 +295,6 @@ type ViewDatabase interface {
 type ColumnOrder struct {
 	First       bool   // True if this column should come first
 	AfterColumn string // Set to the name of the column after which this column should appear
-}
-
-// AlterableTable should be implemented by tables that can receive ALTER TABLE statements to modify their schemas.
-type AlterableTable interface {
-	Table
-	UpdatableTable
-
-	// AddColumn adds a column to this table as given. If non-nil, order specifies where in the schema to add the column.
-	AddColumn(ctx *Context, column *Column, order *ColumnOrder) error
-	// DropColumn drops the column with the name given.
-	DropColumn(ctx *Context, columnName string) error
-	// ModifyColumn modifies the column with the name given, replacing with the new column definition provided (which may
-	// include a name change). If non-nil, order specifies where in the schema to move the column.
-	ModifyColumn(ctx *Context, columnName string, column *Column, order *ColumnOrder) error
 }
 
 // Lockable should be implemented by tables that can be locked and unlocked.
@@ -902,7 +428,7 @@ func IsTrue(val interface{}) bool {
 
 // TypesEqual compares two Types and returns whether they are equivalent.
 func TypesEqual(a, b Type) bool {
-	//TODO: replace all of the Type() == Type() calls with TypesEqual
+	// TODO: replace all of the Type() == Type() calls with TypesEqual
 
 	// We can assume they have the same implementing type if this passes, so we have to check the parameters
 	if a.Type() != b.Type() {
